[refactor]: query함수 커스텀훅으로 리팩토링 ak2j38/issue-tracker#31
